requirements.txt
Pillow
watchdog
customtkinter